% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandFunctions.R
\name{eLag}
\alias{eLag}
\title{Convert vector into a matrix of lag columns}
\usage{
eLag(x, colParamVector, pad = NA)
}
\arguments{
\item{x}{Data \emph{vector}}

\item{colParamVector}{Vector of lags for embedding}

\item{pad}{Scalar for padding embedding}
}
\value{
A matrix whose columns are x lagged by the
        corresponding values in colParamVector.
}
\description{
Convert vector into a matrix of lag columns
}
\examples{
eLag(1:6, 0:2)
eLag(1:6, 0:2, pad=0)
}
\seealso{
\code{\link[stats]{embed}} and
\code{\link[tseriesChaos]{embedd}}, which
 are related functions.
}

