% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandFunctions.R
\name{ePow}
\alias{ePow}
\title{Convert vector x into a matrix \eqn{X_{ij} = {x_i}^j}}
\usage{
ePow(x, colParamVector)
}
\arguments{
\item{x}{Data vector.}

\item{colParamVector}{Vector of column powers.}
}
\value{
A matrix X of size length(x) X length(colParamVector)
        \deqn{X_{ij} = {x_i}^j}
}
\description{
Convert vector x into a matrix \eqn{X_{ij} = {x_i}^j}
}
\examples{
x <- 1:6
ePow(x,0:2)
}

