% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandFunctions.R
\name{freemanTukey}
\alias{freemanTukey}
\title{Freeman-Tukey transform}
\usage{
freemanTukey(x)
}
\arguments{
\item{x}{Data vector}
}
\value{
The transformed vector
}
\description{
This transform takes Poisson (count) information and
makes it more Gaussian, then z-scales (standardizes
by centering and scaling to var = 1) the results.
}
\examples{
x <- freemanTukey(sunspot.month)
}
\references{
Taken from
\url{https://en.wikipedia.org/wiki/Anscombe_transform}
}

