% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandFunctions.R
\name{reset.warnings}
\alias{reset.warnings}
\title{Reset annoyingly persistent warning messages.}
\usage{
reset.warnings()
}
\value{
Returns TRUE invisibly.  Used for side effects only.
}
\description{
Reset annoyingly persistent warning messages.
}
\examples{
## reset.warnings()
}
\references{
This function is built around the snippet found here:
\url{http://stackoverflow.com/questions/5725106/r-how-to-clear-all-warnings}
}

